backend/requirements.txt
django<5
djangorestframework
djongo
pymongo
pydicom
django-cors-headers
